Refactor: Compartmentalize application logic
- I created a `components` directory and moved page components (`HomePage`, `VehicleForm`, `AdminPage`, `VehicleDetailPage`, `DigitalCredential`, `DashboardPage`) into `src/components/pages/`.
- I created a `services` directory for Firebase and other external service interactions:
  - `firebase.js` for Firebase app initialization.
  - `firestoreService.js` for Firestore database operations.
  - `geminiService.js` for Google Gemini API calls.
- I created a `hooks` directory for custom React hooks:
  - `useSnackbar.js` for managing Snackbar notifications.
  - `useAuth.js` for handling Firebase authentication state.
- I created a `theme` directory for Material-UI theme configuration and related constants.
- I refactored `src/App.js` to act as the main application coordinator, utilizing the new components, services, and hooks.
- I updated all import paths across the application to reflect the new structure.

This refactoring improves code organization, modularity, and maintainability.
