+++ src/hooks/useSnackbar.js
@@ -1,0 +1,18 @@
+import { useState } from 'react';
+
+export const useSnackbar = () => {
+    const [snackbar, setSnackbar] = useState({ open: false, message: '', severity: 'success' });
+
+    const showSnackbar = (message, severity = 'success') => {
+        setSnackbar({ open: true, message, severity });
+    };
+
+    const handleCloseSnackbar = (_, reason) => {
+        if (reason === 'clickaway') {
+            return;
+        }
+        setSnackbar({ ...snackbar, open: false });
+    };
+
+    return { snackbar, showSnackbar, handleCloseSnackbar };
+};
